Delete all references to MongoDB broker
We don't use this broker and it is likely broken since all the changes
to this release were made. We also don't see a need for this broker in
the near future since we know how many MongoDB instances we want to
deploy and are capable/prefer to manage them directly with BOSH rather
than through a broker.

Delete all references to the broker and smoke-tests/example apps for
testing the broker

Co-authored-by: Subhash <subhash.thankathurai@smarsh.com>
Co-authored-by: Vin <vinayak.agarwal@smarsh.com>